move datetime DBFunctions to functions namespace
