+++ db/types/custom/datetime.py
@@ -1,12 +1,10 @@
-from sqlalchemy import case, func, and_, TEXT
+from sqlalchemy import case, func, and_
 from sqlalchemy.dialects.postgresql import DATE as SA_DATE
 from sqlalchemy.dialects.postgresql import INTERVAL
 from sqlalchemy.dialects.postgresql import TIME as SA_TIME
 from sqlalchemy.dialects.postgresql import TIMESTAMP as SA_TIMESTAMP
 from sqlalchemy.types import TypeDecorator
 
-from db.functions import hints as db_hints
-from db.functions.base import DBFunction, sa_call_sql_function
 from db.types.exceptions import InvalidTypeParameters
 
 
@@ -255,68 +253,4 @@ class Interval(TypeDecorator):
                 func.date_part('seconds', col),
                 'S',
             )
-        )
-
-
-class TruncateToYear(DBFunction):
-    id = 'truncate_to_year'
-    name = 'Truncate to Year'
-    hints = tuple([db_hints.parameter_count(1)])  # TODO extend hints
-
-    @staticmethod
-    def to_sa_expression(col):
-        return sa_call_sql_function('to_char', col, 'YYYY', return_type=TEXT)
-
-
-class TruncateToMonth(DBFunction):
-    id = 'truncate_to_month'
-    name = 'Truncate to Month'
-    hints = tuple([db_hints.parameter_count(1)])  # TODO extend hints
-
-    @staticmethod
-    def to_sa_expression(col):
-        return sa_call_sql_function('to_char', col, 'YYYY-MM', return_type=TEXT)
-
-
-class TruncateToDay(DBFunction):
-    id = 'truncate_to_day'
-    name = 'Truncate to Day'
-    hints = tuple([db_hints.parameter_count(1)])  # TODO extend hints
-
-    @staticmethod
-    def to_sa_expression(col):
-        return sa_call_sql_function('to_char', col, 'YYYY-MM-DD', return_type=TEXT)
-
-
-class CurrentDate(DBFunction):
-    id = 'current_date'
-    name = 'current date'
-    hints = tuple([db_hints.returns(db_hints.date), db_hints.parameter_count(0)])
-
-    @staticmethod
-    def to_sa_expression():
-        return sa_call_sql_function('current_date', return_type=DATE)
-
-
-class CurrentTime(DBFunction):
-    id = 'current_time'
-    name = 'current time'
-    hints = tuple([db_hints.returns(db_hints.time), db_hints.parameter_count(0)])
-
-    @staticmethod
-    def to_sa_expression():
-        return sa_call_sql_function(
-            'current_time', return_type=TIME_WITH_TIME_ZONE
-        )
-
-
-class CurrentDateTime(DBFunction):
-    id = 'current_datetime'
-    name = 'current datetime'
-    hints = tuple([db_hints.returns(db_hints.date, db_hints.time), db_hints.parameter_count(0)])
-
-    @staticmethod
-    def to_sa_expression():
-        return sa_call_sql_function(
-            'current_timestamp', return_type=TIMESTAMP_WITH_TIME_ZONE
         )
